@@ -1,43 +1,35 @@
-buildscript {
-    repositories {
-        jcenter()
-
-    }
-    dependencies {
-        // Check for the latest version here: http://plugins.gradle.org/plugin/com.jfrog.artifactory
-        classpath "org.jfrog.buildinfo:build-info-extractor-gradle:4+"
-    }
+plugins {
+    id 'java'
+    id 'maven-publish'
 }
 
-allprojects {
-    apply plugin: "com.jfrog.artifactory"
-}
-
-apply plugin: 'java'
-apply plugin: 'application'
-apply plugin: 'maven-publish'
-
-sourceCompatibility = 1.8
-targetCompatibility = 1.8
+sourceCompatibility = 16
+targetCompatibility = 16
 
 def libraryGroupId = 'com.reynke.sloud'
-def libraryArtifactId = 'core'
-def libraryVersion = '1.0.1'
+def libraryArtifactId = rootProject.name
+def libraryVersion = '1.0.2'
 def libraryMainClassName = 'CorePlugin'
 
-group libraryGroupId + '.' + libraryArtifactId
+group libraryGroupId
 version libraryVersion
-mainClassName = group + '.' + libraryMainClassName
+
+def mainClassName = group + '.' + libraryArtifactId + '.' + libraryMainClassName
 
 repositories {
     mavenCentral()
 
     maven {
-        url 'https://hub.spigotmc.org/nexus/content/repositories/snapshots/'
+        url 'https://papermc.io/repo/repository/maven-public/'
     }
 
     maven {
-        url 'https://oss.sonatype.org/content/repositories/snapshots/'
+        name = "GitHubPackages"
+        url = uri("https://maven.pkg.github.com/sloud/database-utilities")
+        credentials {
+            username = project.findProperty("gpr.user") ?: System.getenv("USERNAME")
+            password = project.findProperty("gpr.key") ?: System.getenv("TOKEN")
+        }
     }
 }
 
